Make script wait for MQTT to be up
Often times when HA is restarted, the ebusd addon starts before the
MQTT addon and obviously fails because MQTT is not yet up. This commit
allows some tolerance here by waiting up to 60 seconds for MQTT TCP
connectivity to be possible. It'll try starting ebusd even if this check
fails.