Allow Logging to file
Since ebusd is running in foregraound, the `--logfile` option does not
work. There is however a fairly easy workaround of redirecting the
output to s6-log that can maintain the output to stdout as well as
sending it to a defined folder. The default setting of s6-log is
that it maintains up to 5 files, 1MB each.

This option is especially useful for debugging.